requirements.txt
numpy>=1.20.0
pyyaml
cryptography
rich
psutil
toml
networkx
klayout==0.29.11
z3-solver
matplotlib>=3.5.0
pysmt
python-gdsii
pretty-errors
shapely
tqdm
pandas>=1.3.0
yaspin
scikit-learn == 1.3.2 
tensorflow == 2.15.0
joblib
cryptography
rich
psutil
pyyaml
yaspin
sqlalchemy==1.4.16
torch == 2.3.1
boto3
grpcio
grpcio-tools
python-dotenv==1.0.1
torch_geometric==2.3.0
optuna==4.0.0
optuna-integration==4.0.0
mlflow==2.17.2
fpdf2==2.8.1
pyinstaller>=5.0.0
scipy>=1.7.0
Pillow>=8.0.0